Moved redirectPostToGet middleware from CountryController to its own file in middlewares directory.
